+++ src/views/pages/home/CategoryHome.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react'
+
+export const CategoryHome: FC = () => {
+  return <p>カテゴリー</p>
+}
