+++ src/views/pages/home/LikeHome.tsx
@@ -1,0 +1,5 @@
+import { FC } from 'react'
+
+export const LikeHome: FC = () => {
+  return <p>お気に入り</p>
+}
